+++ src/main/java/com/fifteen/auction/global/client/chatgpt/ChatGPTClient.java
@@ -1,6 +1,6 @@
-package com.fifteen.auction.global.client;
+package com.fifteen.auction.global.client.chatgpt;
 
-import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fifteen.auction.domain.product.dto.response.GPTPricePredictionResponse;
 import com.fifteen.auction.global.client.chatgpt.dto.ChatGPTRequest;
